+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Unreleased changes
 
 ### What Changed
 
+* Added support for template attributes like CKA_DERIVE_TEMPLATE, which can
+  be used to enforce specific attribute on key create through use of other keys
+  (Derivation, encapsualtion, wrapping)
+  - [Template attributes](https://github.com/latchset/kryoptic/pull/500)
 
 
 ## [1.5.1]